upgrade greptime-proto
