+++ src/meta-srv/src/selector/load_based.rs
@@ -17,7 +17,8 @@ use common_telemetry::warn;
 
 use crate::cluster::MetaPeerClient;
 use crate::error::Result;
-use crate::keys::{LeaseKey, LeaseValue, StatKey};
+use crate::handler::node_stat::RegionStat;
+use crate::keys::{LeaseKey, LeaseValue, StatKey, StatValue};
 use crate::lease;
 use crate::metasrv::SelectorContext;
 use crate::selector::{Namespace, Selector};
@@ -45,12 +46,13 @@ impl Selector for LoadBasedSelector {
 
         let mut tuples: Vec<(LeaseKey, LeaseValue, u64)> = lease_kvs
             .into_iter()
+            // The regions of a table need to be distributed on different datanode.
             .filter(|(lease_k, _)| {
                 if let Some(stat_val) = stat_kvs.get(&lease_k.into()) {
                     if let (Some(catalog), Some(schema), Some(table)) =
                         (&ctx.catalog, &ctx.schema, &ctx.table)
                     {
-                        return stat_val.contains_table(catalog, schema, table) != Some(true);
+                        return contains_table(stat_val, catalog, schema, table) != Some(true);
                     }
                 }
                 true
@@ -83,5 +85,129 @@ impl Selector for LoadBasedSelector {
                 addr: lease_val.node_addr,
             })
             .collect())
+    }
+}
+
+// Determine whether there is a table in dn according to the heartbeats.
+//
+// Result:
+// None indicates no heartbeats in stat_val;
+// Some(true) indicates table exists in the datanode;
+// Some(false) indicates that table not exists in datanode.
+fn contains_table(
+    stat_val: &StatValue,
+    catalog_name: &str,
+    schema_name: &str,
+    table_name: &str,
+) -> Option<bool> {
+    let may_latest = stat_val.stats.last();
+
+    if let Some(latest) = may_latest {
+        for RegionStat {
+            catalog,
+            schema,
+            table,
+            ..
+        } in latest.region_stats.iter()
+        {
+            if catalog_name == catalog && schema_name == schema && table_name == table {
+                return Some(true);
+            }
+        }
+    } else {
+        return None;
+    }
+
+    Some(false)
+}
+
+#[cfg(test)]
+mod tests {
+    use crate::handler::node_stat::{RegionStat, Stat};
+    use crate::keys::StatValue;
+    use crate::selector::load_based::contains_table;
+
+    #[test]
+    fn test_contains_table_from_stat_val() {
+        let empty = StatValue { stats: vec![] };
+        assert!(contains_table(&empty, "greptime_4", "public_4", "demo_5").is_none());
+
+        let stat_val = StatValue {
+            stats: vec![
+                Stat {
+                    region_stats: vec![
+                        RegionStat {
+                            catalog: "greptime_1".to_string(),
+                            schema: "public_1".to_string(),
+                            table: "demo_1".to_string(),
+                            ..Default::default()
+                        },
+                        RegionStat {
+                            catalog: "greptime_2".to_string(),
+                            schema: "public_2".to_string(),
+                            table: "demo_2".to_string(),
+                            ..Default::default()
+                        },
+                        RegionStat {
+                            catalog: "greptime_3".to_string(),
+                            schema: "public_3".to_string(),
+                            table: "demo_3".to_string(),
+                            ..Default::default()
+                        },
+                    ],
+                    ..Default::default()
+                },
+                Stat {
+                    region_stats: vec![
+                        RegionStat {
+                            catalog: "greptime_1".to_string(),
+                            schema: "public_1".to_string(),
+                            table: "demo_1".to_string(),
+                            ..Default::default()
+                        },
+                        RegionStat {
+                            catalog: "greptime_2".to_string(),
+                            schema: "public_2".to_string(),
+                            table: "demo_2".to_string(),
+                            ..Default::default()
+                        },
+                        RegionStat {
+                            catalog: "greptime_3".to_string(),
+                            schema: "public_3".to_string(),
+                            table: "demo_3".to_string(),
+                            ..Default::default()
+                        },
+                    ],
+                    ..Default::default()
+                },
+                Stat {
+                    region_stats: vec![
+                        RegionStat {
+                            catalog: "greptime_1".to_string(),
+                            schema: "public_1".to_string(),
+                            table: "demo_1".to_string(),
+                            ..Default::default()
+                        },
+                        RegionStat {
+                            catalog: "greptime_2".to_string(),
+                            schema: "public_2".to_string(),
+                            table: "demo_2".to_string(),
+                            ..Default::default()
+                        },
+                        RegionStat {
+                            catalog: "greptime_4".to_string(),
+                            schema: "public_4".to_string(),
+                            table: "demo_4".to_string(),
+                            ..Default::default()
+                        },
+                    ],
+                    ..Default::default()
+                },
+            ],
+        };
+        assert!(contains_table(&stat_val, "greptime_1", "public_1", "demo_1").unwrap());
+        assert!(contains_table(&stat_val, "greptime_2", "public_2", "demo_2").unwrap());
+        assert!(!contains_table(&stat_val, "greptime_3", "public_3", "demo_3").unwrap());
+        assert!(contains_table(&stat_val, "greptime_4", "public_4", "demo_4").unwrap());
     }
 }
