+++ src/meta-srv/src/service/router.rs
@@ -138,7 +138,7 @@ async fn handle_create(
         return Ok(RouteResponse {
             header: Some(ResponseHeader::failed(
                 cluster_id,
-                Error::not_enough_active_datanodes(peers.len() as u32),
+                Error::not_enough_available_datanodes(partitions.len(), peers.len()),
             )),
             ..Default::default()
         });
