+++ tests-integration/src/grpc.rs
@@ -109,7 +109,6 @@ mod test {
                 table_name: "table_created_through_grpc".to_string(),
                 kind: Some(alter_expr::Kind::AddColumns(AddColumns {
                     add_columns: vec![AddColumn {
-                        location: None,
                         column_def: Some(ColumnDef {
                             name: "b".to_string(),
                             datatype: ColumnDataType::Int32 as _,
